pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "polymarket-trading"
version = "0.1.0"
requires-python = ">=3.10"
dependencies = [
    "py-clob-client>=0.34",
    "py-builder-relayer-client>=0.0.1",
    "python-dotenv>=1.0",
    "requests>=2.31",
    "websocket-client>=1.7",
    "pandas>=2.1",
    "pydantic>=2.5",
    "betfairlightweight>=2.19",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4",
    "pytest-asyncio>=0.23",
    "ruff>=0.3",
    "mypy>=1.8",
]
